Move: build_entry_path() from imag-store to libimagstore
